+++ src/editor/sagas.ts
@@ -1,14 +1,29 @@
 // SPDX-License-Identifier: MIT
-// Copyright (c) 2020-2021 The Pybricks Authors
+// Copyright (c) 2020-2022 The Pybricks Authors
 
 import FileSaver from 'file-saver';
-import { call, put, select, takeEvery } from 'typed-redux-saga/macro';
+import {
+    call,
+    put,
+    race,
+    select,
+    take,
+    takeEvery,
+    takeLatest,
+} from 'typed-redux-saga/macro';
+import { Action } from '../actions';
+import {
+    FileStorageActionType,
+    FileStorageDidFailToReadFileAction,
+    FileStorageDidReadFileAction,
+    fileStorageReadFile,
+} from '../fileStorage/actions';
 import { RootState } from '../reducers';
 import { ensureError } from '../utils';
 import {
+    CurrentEditorAction,
     EditorActionType,
     EditorOpenAction,
-    EditorReloadProgramAction,
     EditorSaveAsAction,
     didFailToSaveAs,
     didSaveAs,
@@ -78,20 +93,48 @@ function* saveAs(_action: EditorSaveAsAction): Generator {
     yield* put(didSaveAs());
 }
 
-function* reloadProgram(_action: EditorReloadProgramAction): Generator {
-    const editor = yield* select((s: RootState) => s.editor.current);
-
-    // istanbul ignore next: it is a bug to dispatch this action with no current editor
-    if (editor === null) {
-        console.error('reloadProgram: No current editor');
+function* handleEditSession(action: CurrentEditorAction): Generator {
+    if (action.editSession === null) {
+        // there is not current edit session, nothing to do
         return;
     }
 
-    editor.setValue(localStorage.getItem('program') || '');
+    // ensure storage has been initialized
+
+    const isStorageInitialized = yield* select(
+        (s: RootState) => s.fileStorage.isInitialized,
+    );
+
+    if (!isStorageInitialized) {
+        yield* take(FileStorageActionType.DidInitialize);
+    }
+
+    // TODO: get current file from state
+    const currentFileName = 'main.py';
+
+    // TODO: implement locking to ensure exclusive access to file
+
+    yield* put(fileStorageReadFile(currentFileName));
+    const { result } = yield* race({
+        result: take<FileStorageDidReadFileAction>(
+            (a: Action) =>
+                a.type === FileStorageActionType.DidReadFile &&
+                a.fileName === currentFileName,
+        ),
+        error: take<FileStorageDidFailToReadFileAction>(
+            (a: Action) =>
+                a.type === FileStorageActionType.DidFailToReadFile &&
+                a.fileName === currentFileName,
+        ),
+    });
+
+    if (result) {
+        action.editSession?.setValue(result.fileContents);
+    }
 }
 
 export default function* (): Generator {
     yield* takeEvery(EditorActionType.Open, open);
     yield* takeEvery(EditorActionType.SaveAs, saveAs);
-    yield* takeEvery(EditorActionType.ReloadProgram, reloadProgram);
+    yield* takeLatest(EditorActionType.Current, handleEditSession);
 }